+++ backend/src/main/java/turip/data/service/DataImportService.java
@@ -10,24 +10,24 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import turip.place.domain.Category;
-import turip.place.repository.CategoryRepository;
-import turip.region.domain.City;
-import turip.region.repository.CityRepository;
 import turip.content.domain.Content;
-import turip.content.repository.ContentRepository;
 import turip.content.domain.ContentPlace;
 import turip.content.repository.ContentPlaceRepository;
-import turip.region.domain.Country;
-import turip.region.repository.CountryRepository;
+import turip.content.repository.ContentRepository;
 import turip.creator.domain.Creator;
 import turip.creator.repository.CreatorRepository;
 import turip.data.controller.dto.CsvDataDto;
+import turip.place.domain.Category;
 import turip.place.domain.Place;
-import turip.place.repository.PlaceRepository;
 import turip.place.domain.PlaceCategory;
+import turip.place.repository.CategoryRepository;
 import turip.place.repository.PlaceCategoryRepository;
+import turip.place.repository.PlaceRepository;
+import turip.region.domain.City;
+import turip.region.domain.Country;
 import turip.region.domain.Province;
+import turip.region.repository.CityRepository;
+import turip.region.repository.CountryRepository;
 import turip.region.repository.ProvinceRepository;
 
 @Slf4j
