+++ backend/src/test/java/turip/favorite/api/FavoritePlaceApiTest.java
@@ -1,4 +1,4 @@
-package turip.favoriteplace.api;
+package turip.favorite.api;
 
 import static org.hamcrest.Matchers.is;
